+++ packages/cashscript/test/util.test.ts
@@ -6,7 +6,7 @@ import {
   placeholder,
 } from '@cashscript/utils';
 import {
-  scriptToP2sh32Address,
+  scriptToAddress,
   createInputScript,
   getInputSize,
   getPreimageSize,
@@ -81,22 +81,31 @@ describe('utils', () => {
     });
   });
 
-  describe('scriptToP2sh32Address', () => {
+  describe('scriptToAddress', () => {
     it('should convert a redeem script to a cashaddress', () => {
       const asm = `${binToHex(alicePkh)} OP_OVER OP_HASH160 OP_EQUALVERIFY OP_CHECKSIG`;
       const redeemScript = asmToScript(asm);
 
-      const mainnetAddress = scriptToP2sh32Address(redeemScript, Network.MAINNET);
-      const testnetAddress = scriptToP2sh32Address(redeemScript, Network.TESTNET4);
-      const regtestAddress = scriptToP2sh32Address(redeemScript, Network.REGTEST);
+      const mainnetAddressP2sh32 = scriptToAddress(redeemScript, Network.MAINNET, 'p2sh32');
+      const testnetAddressP2sh32 = scriptToAddress(redeemScript, Network.TESTNET4, 'p2sh32');
+      const regtestAddressP2sh32 = scriptToAddress(redeemScript, Network.REGTEST, 'p2sh32');
+      const mainnetAddressP2sh20 = scriptToAddress(redeemScript, Network.MAINNET, 'p2sh20');
+      const testnetAddressP2sh20 = scriptToAddress(redeemScript, Network.TESTNET4, 'p2sh20');
+      const regtestAddressP2sh20 = scriptToAddress(redeemScript, Network.REGTEST, 'p2sh20');
 
-      const expectedMainnetAddress = 'bitcoincash:pv6dnl7ws66dzdk2wn5akmmyx0f4fztx56lqvszjuu52fsw3d23629h20jd0s';
-      const expectedTestnetAddress = 'bchtest:pv6dnl7ws66dzdk2wn5akmmyx0f4fztx56lqvszjuu52fsw3d2362xsm3276z';
-      const expectedRegtestAddress = 'bchreg:pv6dnl7ws66dzdk2wn5akmmyx0f4fztx56lqvszjuu52fsw3d2362n6fmdfd2';
+      const expectedMainnetAddressP2sh32 = 'bitcoincash:pv6dnl7ws66dzdk2wn5akmmyx0f4fztx56lqvszjuu52fsw3d23629h20jd0s';
+      const expectedTestnetAddressP2sh32 = 'bchtest:pv6dnl7ws66dzdk2wn5akmmyx0f4fztx56lqvszjuu52fsw3d2362xsm3276z';
+      const expectedRegtestAddressP2sh32 = 'bchreg:pv6dnl7ws66dzdk2wn5akmmyx0f4fztx56lqvszjuu52fsw3d2362n6fmdfd2';
+      const expectedMainnetAddressP2sh20 = 'bitcoincash:pr4wzdh0h9d7fpu890lq8xz0c84cpv3nvyc27hzc7y';
+      const expectedTestnetAddressP2sh20 = 'bchtest:pr4wzdh0h9d7fpu890lq8xz0c84cpv3nvyuc6sq0ec';
+      const expectedRegtestAddressP2sh20 = 'bchreg:pr4wzdh0h9d7fpu890lq8xz0c84cpv3nvyxyv3ru67';
 
-      expect(mainnetAddress).toEqual(expectedMainnetAddress);
-      expect(testnetAddress).toEqual(expectedTestnetAddress);
-      expect(regtestAddress).toEqual(expectedRegtestAddress);
+      expect(mainnetAddressP2sh32).toEqual(expectedMainnetAddressP2sh32);
+      expect(testnetAddressP2sh32).toEqual(expectedTestnetAddressP2sh32);
+      expect(regtestAddressP2sh32).toEqual(expectedRegtestAddressP2sh32);
+      expect(mainnetAddressP2sh20).toEqual(expectedMainnetAddressP2sh20);
+      expect(testnetAddressP2sh20).toEqual(expectedTestnetAddressP2sh20);
+      expect(regtestAddressP2sh20).toEqual(expectedRegtestAddressP2sh20);
     });
   });
 });